requirements.txt
appdirs==1.4.4
asgiref==3.4.0
autopep8==1.5.7
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
distlib==0.3.2
dj-database-url==0.5.0
Django==3.2.4
django-crispy-forms==1.12.0
django-heroku==0.3.1
filelock==3.0.12
gunicorn==20.1.0
idna==2.10
Pillow==8.3.0
psycopg2==2.9.1
pycodestyle==2.7.0
pytz==2021.1
requests==2.25.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
tmdbv3api==1.7.5
toml==0.10.2
urllib3==1.26.6
virtualenv==20.4.7
whitenoise==5.2.0